Hello world SC visual Synth

+++ src/ofxJVisuals.hpp
@@ -96,8 +96,8 @@ typedef unsigned char uint8;
     void getFreePointers(string ip, int port);
     unsigned short numEventsAdded = 0;
 
-    JEvent* addEvent(JEvent* e, int layerIndex=1, unsigned short index = 0);
-    JEvent* addEvent(JEvent* e, VisualizerLayer l, unsigned short index = 0);
+    JEvent* addEvent(JEvent* e, int layerIndex=1, int index = 0);
+    JEvent* addEvent(JEvent* e, VisualizerLayer l, int index = 0);
 
     JEvent* getLast(){return last;}
     JEvent* getEventById(int id);
